ENH: Improve logging in getDICOMValue function
Do not log error message if a tag is not found in the DICOM data set, as it is not an error case (that's why the default value is provided). We don't log an error when tag is retrieved from the database either.

If some unknown error happens and an exception is thrown then log the full traceback to make troubleshooting easier.